+++ tui/stopwatch/components/laps/init.go
@@ -1,0 +1,10 @@
+package laps
+
+import tea "github.com/charmbracelet/bubbletea"
+
+// INIT
+// ----
+
+func (m Model) Init() tea.Cmd {
+	return nil
+}
