+++ tui/stopwatch/components/laps/model.go
@@ -1,14 +1,6 @@
 package laps
 
-import (
-	"fmt"
-	"time"
-
-	tea "github.com/charmbracelet/bubbletea"
-	"github.com/charmbracelet/lipgloss"
-
-	"github.com/Shresht7/go-time/tui/helpers"
-)
+import "time"
 
 // MODEL
 // -----
@@ -75,44 +67,4 @@ func (m Model) sum(i int) time.Duration {
 		total += lap.duration()
 	}
 	return total
-}
-
-// INIT
-// ----
-
-func (m Model) Init() tea.Cmd {
-	return nil
-}
-
-// UPDATE
-// ------
-
-func (m Model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
-	return m, nil
-}
-
-// VIEW
-// ----
-
-func (m Model) View() string {
-	s := ""
-	for i, lap := range m.laps {
-		s += fmt.Sprintf("%s\t%s\t\t%s\n", m.styleIndex(i), helpers.FormatDuration(lap.duration()), helpers.FormatDuration(m.sum(i)))
-	}
-	return s
-}
-
-var colorRed = lipgloss.NewStyle().Foreground(lipgloss.Color("9"))
-var colorBlue = lipgloss.NewStyle().Foreground(lipgloss.Color("12"))
-
-func (m Model) styleIndex(i int) string {
-	s := fmt.Sprintf("%d", i+1)
-
-	if i == m.Shortest() {
-		return colorBlue.Render(s)
-	}
-	if i == m.Longest() {
-		return colorRed.Render(s)
-	}
-	return s
 }
