+++ tui/stopwatch/components/laps/update.go
@@ -1,0 +1,10 @@
+package laps
+
+import tea "github.com/charmbracelet/bubbletea"
+
+// UPDATE
+// ------
+
+func (m Model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
+	return m, nil
+}
